requirements.txt
pytest~=8.3.4
loguru~=0.7.3
playwright~=1.49.1
pip~=25.0.1
attrs~=24.3.0
wheel~=0.41.2
allure-python-commons~=2.13.5